SPC: Run SPC 1 frame per frame, rather than 60 frames per frame. + Fixed warnings in SPC code in 64-bit mode

+++ Core/SPC_DSP.h
@@ -154,7 +154,7 @@ private:
 
 #include <assert.h>
 
-inline int SPC_DSP::sample_count() const { return m.out - m.out_begin; }
+inline int SPC_DSP::sample_count() const { return (int)(m.out - m.out_begin); }
 
 inline int SPC_DSP::read( int addr ) const
 {
